// src/settings/type.ts
import { App, Menu, MenuItem } from "obsidian";

export type EnhancedMenu = Menu & {
  dom: HTMLElement;
  items: EnhancedMenuItem[];
  setUseNativeMenu: Function;
};

export type EnhancedMenuItem = MenuItem & {
  dom: HTMLElement;
  titleEl: HTMLDivElement;
  handleEvent(event: Event): void;
  disabled: boolean;
};

export type EnhancedApp = App & {
  customCss: {
    snippets: string[];
    getSnippetsFolder: Function;
    getSnippetPath: Function;
    enabledSnippets: Set<string>;
    setCssEnabledStatus: Function;
    requestLoadSnippets: Function;
  };
  openWithDefaultApp: Function;
};